Adding time slicing of 10 second parts to the recording of group streams received via WebRTC, the reassembly logic is done by the admin to better handle out of order parts

+++ gwt-cordova/src/main/java/nl/mpi/tg/eg/experiment/client/listener/MediaSubmissionListener.java
@@ -46,7 +46,7 @@ public abstract class MediaSubmissionListener {
 
     public abstract void recorderStarted(final String targetDeviceId, final Double audioContextCurrentMS);
 
-    public abstract void submissionFailed(final String message, final Uint8Array dataArray, final Integer partNumber);
+    public abstract void submissionFailed(final String message, final Uint8Array dataArray, final Double partNumber);
 
     public abstract void submissionComplete(String message);
